refactor(step3): 패키지 수정, optimize imports
