+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -12,9 +12,9 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToOne;
 
-import kitchenpos.common.price.domain.Price;
 import kitchenpos.common.exception.KitchenposErrorCode;
 import kitchenpos.common.exception.KitchenposException;
+import kitchenpos.common.price.domain.Price;
 import kitchenpos.menugroup.domain.MenuGroup;
 
 @Entity
