+++ src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -4,8 +4,6 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.menu.domain.Menu;
-
 public interface MenuRepository extends JpaRepository<Menu, Long> {
     long countByIdIn(List<Long> ids);
 }
