+++ src/main/java/kitchenpos/order/domain/OrderLineItemRepository.java
@@ -1,8 +1,6 @@
-package kitchenpos.order.repository;
+package kitchenpos.order.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import kitchenpos.order.domain.OrderLineItem;
 
 public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
 }
