+++ src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -1,11 +1,9 @@
-package kitchenpos.order.repository;
+package kitchenpos.order.domain;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
