+++ src/main/java/kitchenpos/tablegroup/dto/TableGroupResponse.java
@@ -3,9 +3,9 @@ package kitchenpos.tablegroup.dto;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.table.dto.OrderTableResponse;
 import kitchenpos.table.dto.OrderTableResponses;
+import kitchenpos.tablegroup.domain.TableGroup;
 
 public class TableGroupResponse {
     private Long id;
