+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -28,7 +28,7 @@ import kitchenpos.order.dto.OrderResponse;
 import kitchenpos.common.exception.KitchenposException;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.order.repository.OrderRepository;
+import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
